Major reorganization of the repo
Reorganized the repo for packaging for mise 0.0.1. Source code moved
to mise/src/mise which required changing class calls. Old source files
were at mise/mise and so if there is a history of a file missing it can
be found in the git history there. added build dependencies with
pyproject.toml as well as packaging scripts in /scripts dir for
creating packages on macOS and Windows.