Refactored ProjectWidget and dependencies
Before developing the analysis view it became necessary to split off
the code tree and file tree widgets from where they were originally
instantiated. This has necessaitated refactoring further and creating
new signals and slots. I have not done a very disciplined job of
completing it and the app is working but I worry that it is very
fragile at this point. This commit should be considered canonical in
the sense that it is a fall back for future work.