Refactoring to fix use of paths
v0.1.0 was built with classes and methods using absolut paths to
project files which is not a problem until the project directory
moves and the text path stored in the database changes. A number
of classes and methods used paths to pass information between them.
I have refactored in a very fast and loose way to plug these issues
but it will require a more substantial refactoring to change the
architecture of Mise to rely upon the metadata stored in the database
to return the relative path to text documents.